backend/requirements.txt
py-algorand-sdk>=2.0.0
pyteal>=0.24.0
flask>=3.0
flask-cors>=4.0
flask-jwt-extended>=4.6
python-dotenv>=1.0
requests>=2.31
